+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/verticalmode/PaymentMethodEmbeddedLayoutUIScreenshotTest.kt
@@ -42,7 +42,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
@@ -62,7 +61,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
@@ -82,7 +80,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
@@ -102,7 +99,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
@@ -127,7 +123,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
@@ -151,7 +146,6 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 PaymentMethodVerticalLayoutInteractor.SavedPaymentMethodAction.MANAGE_ALL,
                 selection = PaymentSelection.Saved(savedPaymentMethod.paymentMethod),
                 isEnabled = true,
-                onEditPaymentMethod = {},
                 onViewMorePaymentMethods = {},
                 onSelectSavedPaymentMethod = {},
                 onManageOneSavedPaymentMethod = {},
